Move client/ to sdks/go/ for multi-language SDK layout
Relocate the Go client SDK from client/ to sdks/go/ so future SDKs in
other languages (Python, TypeScript, ...) can live as siblings under
sdks/. The Go module stays single-rooted at github.com/anaregdesign/lantern
because the SDK still depends on core/graph; splitting into a separate
module would require replace directives or go.work for trivial gain at
this stage.

Changes:
- git mv client sdks/go (preserves history on all 8 files)
- Rewrite import path github.com/anaregdesign/lantern/client ->
  github.com/anaregdesign/lantern/sdks/go in cli/, server/service/
  integration_test.go, sdks/go/example, sdks/go/client_options_test.go,
  and README.md
- Update AGENTS.md and README.md to point at the new location and note
  the path change in the repo layout tables
- Package name remains 'client', so user code reads:
    import "github.com/anaregdesign/lantern/sdks/go"
    cli, _ := client.NewLantern(...)

go build ./... and go test ./... both pass.
