+++ sdks/go/client_test.go
