+++ sdks/go/value_kind_test.go
